+++ packages/sphinx-autodoc-fastmcp/src/sphinx_autodoc_fastmcp/_collector.py
@@ -8,7 +8,7 @@ import logging
 import typing as t
 
 from sphinx.application import Sphinx
-from sphinx_typehints_gp import normalize_annotation_text
+from sphinx_autodoc_typehints_gp import normalize_annotation_text
 
 from sphinx_autodoc_fastmcp._models import ToolInfo
 from sphinx_autodoc_fastmcp._parsing import extract_params