+++ packages/sphinx-autodoc-pytest-fixtures/src/sphinx_autodoc_pytest_fixtures/__init__.py
@@ -100,7 +100,7 @@ def setup(app: Sphinx) -> SetupDict:
     app.setup_extension("sphinx.ext.autodoc")
     app.setup_extension("sphinx_autodoc_badges")
     app.setup_extension("sphinx_autodoc_layout")
-    app.setup_extension("sphinx_typehints_gp")
+    app.setup_extension("sphinx_autodoc_typehints_gp")
 
     import pathlib