+++ tests/ext/typehints_gp/__init__.py
@@ -1,1 +1,1 @@
-"""Tests for sphinx_typehints_gp."""
+"""Tests for sphinx_autodoc_typehints_gp."""